+++ src/main/groovy/org/gradle/api/plugins/cargo/LocalContainerTask.groovy
@@ -42,6 +42,16 @@ class LocalContainerTask extends AbstractContainerTask {
     void validateConfiguration() {
         super.validateConfiguration()
 
+        getDeployables().each { deployable ->
+            if(deployable.file && !deployable.file.exists()) {
+                throw new InvalidUserDataException("Deployable "
+                        + (deployable.file == null ? "null" : deployable.file.canonicalPath)
+                        + " does not exist")
+            }
+
+            logger.info "Deployable artifacts = ${getDeployables().collect { it.file.canonicalPath }}"
+        }
+
         if(!getConfigFiles().isEmpty()) {
             getConfigFiles().each { configFile ->
                 if(!configFile.file || !configFile.file.exists()) {